# src/infra/rest-api/endpoints/openapi.yaml
openapi: 3.1.0
info:
  title: sump api
  description: sump - simple user management platform
  version: 0.1.0
...
components:
  schemas:
...
    CreateNewAccountOperationDto:
      type: object
      required:
        - name
        - username
        - email
      properties:
        name:
          type: string
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        roles:
          type: array
          items:
            type: string
...
    CreateNewEnvironmentAccountOperationDto:
      type: object
      required:
        - name
        - customProperties
      properties:
        name:
          type: string
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        customProperties:
          type: object
...
    UpdateAccountOperationDto:
      type: object
      properties:
        name:
          type: string
    UpdateAccountEmailOperationDto:
      type: object
      properties:
        email:
          type: string
    UpdateAccountPhoneOperationDto:
      type: object
      properties:
        phone:
          type: string
...
    UpdateEnvironmentAccountEmailOperationDto:
      type: object
      properties:
        email:
          type: string
    UpdateEnvironmentAccountPhoneOperationDto:
      type: object
      properties:
        phone:
          type: string
...
    SetEnvironmentAccountCustomPropertyOperationDto:
      type: object
    DeleteEnvironmentAccountCustomPropertyOperationDto:
      type: object
...
    UpdateEnvironmentAccountByIdAndEnvironmentIdOperationDto:
      type: object
      properties:
        name:
          type: string
...
    CreateNewAccountOperationDtoResponse:
      type: object
      properties:
        accountId:
          type: string
...
    CreateNewEnvironmentAccountOperationDtoResponse:
      type: object
      properties:
        accountId:
          type: string
...
    GetAccountsByTenantIdOperationDtoResponse:
      type: array
      items:
        type: object
        properties:
          id:
            type: string
          tenantId:
            type: string
          name:
            type: string
          email:
            type: string
          phone:
            type: string
          username:
            type: string
          avatarUrl:
            type: string
          roles:
            type: array
            items:
              type: string
    GetAccountByIdAndTenantIdOperationDtoResponse:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        username:
          type: string
        avatarUrl:
          type: string
        roles:
          type: array
          items:
            type: string
...
    GetEnvironmentAccountByIdAndEnvironmentIdOperationDtoResponse:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        username:
          type: string
        avatarUrl:
          type: string
        customProperties:
          type: object
paths:
...
  /api/v1/tenants/{tenantId}/accounts:
    get:
      operationId: getaccountsbytenantid
      tags:
        - Account
      parameters:
        - name: tenantId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Expected response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAccountsByTenantIdOperationDtoResponse'
    post:
      operationId: createnewaccount
      tags:
        - Account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNewAccountOperationDto'
      responses:
        '201':
          description: 'Expected response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateNewAccountOperationDtoResponse'
  /api/v1/tenants/{tenantId}/accounts/{accountId}:
    get:
      operationId: getaccountbyidandtenantid
      tags:
        - Account
      parameters:
        - name: tenantId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Expected response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAccountByIdAndTenantIdOperationDtoResponse'
    delete:
      operationId: deleteaccountbyidandtenantid
      tags:
        - Account
      parameters:
        - name: tenantId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'Expected response'
    patch:
      operationId: updateaccountbyidandtenantid
      tags:
        - Account
      parameters:
        - name: tenantId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountOperationDto'
      responses:
        '200':
          description: 'Expected response'
  /api/v1/tenants/{tenantId}/accounts/{accountId}/email:
    patch:
      operationId: updateaccountemailbyidandtenantid
      tags:
        - Account
      parameters:
        - name: tenantId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountEmailOperationDto'
      responses:
        '200':
          description: 'Expected response'
  /api/v1/tenants/{tenantId}/accounts/{accountId}/username:
    patch:
      operationId: updateaccountusernamebyidandtenantid
      tags:
        - Account
      parameters:
        - name: tenantId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccounUsernameOperationDto'
      responses:
        '200':
          description: 'Expected response'
  /api/v1/tenants/{tenantId}/accounts/{accountId}/phone:
    patch:
      operationId: updateaccountphonebyidandtenantid
      tags:
        - Account
      parameters:
        - name: tenantId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountPhoneOperationDto'
      responses:
        '200':
          description: 'Expected response'
...
  /api/v1/tenant-environments/{environmentId}/accounts:
    post:
      operationId: createnewenvironmentaccount
      tags:
        - Environment Account
      parameters:
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNewEnvironmentAccountOperationDto'
      responses:
        '201':
          description: 'Expected response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateNewEnvironmentAccountOperationDtoResponse'
  /api/v1/tenant-environments/{environmentId}/accounts/{accountId}:
    patch:
      operationId: updateenvironmentaccountbyidandenvironmentid
      tags:
        - Environment Account
      parameters:
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEnvironmentAccountByIdAndEnvironmentIdOperationDto'
      responses:
        '200':
          description: 'Expected response'
    delete:
      operationId: deleteenvironmentaccountbyidandenvironmentid
      tags:
        - Environment Account
      parameters:
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Expected response'
    get:
      operationId: getnvironmentaccountbyidandenvironmentid
      tags:
        - Environment Account
      parameters:
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Expected response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetEnvironmentAccountByIdAndEnvironmentIdOperationDtoResponse'
  /api/v1/tenant-environments/{environmentId}/accounts/{accountId}/email:
    patch:
      operationId: updateenvironmentaccountemailbyidandtenantid
      tags:
        - Environment Account
      parameters:
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEnvironmentAccountEmailOperationDto'
      responses:
        '200':
          description: 'Expected response'
  /api/v1/tenant-environments/{environmentId}/accounts/{accountId}/username:
    patch:
      operationId: updateenvironmentaccountusernamebyidandtenantid
      tags:
        - Environment Account
      parameters:
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEnvironmentAccounUsernameOperationDto'
      responses:
        '200':
          description: 'Expected response'
  /api/v1/tenant-environments/{environmentId}/accounts/{accountId}/phone:
    patch:
      operationId: updateenvironmentaccountphonebyidandtenantid
      tags:
        - Environment Account
      parameters:
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEnvironmentAccountPhoneOperationDto'
      responses:
        '200':
          description: 'Expected response'
  /api/v1/tenant-environments/{environmentId}/accounts/{accountId}/custom-property:
    patch:
      operationId: updateenvironmentaccountcustompropertiesbyidandtenantid
      tags:
        - Environment Account
      parameters:
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetEnvironmentAccountCustomPropertyOperationDto'
      responses:
        '200':
          description: 'Expected response'

    delete:
      operationId: deleteenvironmentaccountcustompropertiesbyidandtenantid
      tags:
        - Environment Account
      parameters:
        - name: environmentId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
        - name: accountId
          in: path
          description: ID of tenant the accounts will be listed from
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteEnvironmentAccountCustomPropertyOperationDto'
      responses:
        '200':
          description: 'Expected response'